Switched to FastAPI and re-organized server.py.
