+++ applicationInterface/cursor.py
@@ -1,0 +1,16 @@
+import psycopg2
+
+# Create database connection.
+conn = psycopg2.connect(
+    host="localhost",
+    database="commons",
+    user="commons_dev",
+    password="commons_dev"
+)
+"""
+user="commons_dev",
+password="commons_dev"
+"""
+
+# Create cursor to interact with the database.
+cur = conn.cursor()
